requirements.txt
Django==5.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1
python-decouple==3.8
gunicorn==21.2.0
dj-database-url==2.1.0
drf-yasg==1.21.7
requests==2.31.0
